requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
idna==2.10
importlib-metadata==4.5.0
joblib==1.0.1
nltk==3.6.2
PyPDF2==1.26.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
sqlparse==0.4.1
tika==1.24
tqdm==4.61.1
typing-extensions==3.10.0.0
urllib3==1.26.5
zipp==3.4.1
stanza==1.2.2
spacy-legacy==3.0.6

